cloudFunctions/aws/githubDefaultBranchProtection/requirements.txt
boto3
httpx
python-dotenv